Allow disabling or setting a custom GZipMiddleware (#5582)
### Motivation

At the moment there is no way to disable or customize the
GZipMiddleware. In some situations, like for our [Learning
Loop](https://zauberzeug.com/products/learning-loop), this can be
disadvantageous.

### Implementation

This PR adds a parameter to `ui.run`, defined as `gzip_middleware:
type[GZipMiddleware] | None = GZipMiddleware`. This allows the user to
disable the middleware by setting `gzip_middleware` to `None`, or to
configure it via a custom subclass.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [ ] Documentation has been added (necessary?).

---------

Co-authored-by: Dr. Dennis Wittich <dennis@zauberzeug.com>