Bake `ajv-formats` into `ui.json_editor` (#5571)
### Motivation

This PR is in response to #4748 in which UUID schema is not supported.
This requires `ajv-formats`, which this PR _nicely_ packs into NiceGUI
as a permanent inclusion

### Implementation

Contrast to #5559 this is a very "batteries-included" PR, to the point
it is not possible to opt-out of `ajv-formats`.

It is done this way because the size increase isn't that large (+10%
only) and we can do so much more with `ajv-formats` installed.

I can tell you, the trick to install `@rollup/plugin-commonjs` and
`@rollup/plugin-json` foils both ChatGPT, Grok, and Claude.

By manually putting in the work, I prove that [I do not have skill
issue](https://github.com/zauberzeug/nicegui/issues/4748#issuecomment-3628309903).

### Demo code

```py
schema = {
    'type': 'object',
    'properties': {
        'uuid': {
            'type': 'string',
            'format': 'uuid',  # breaks on main, not on this PR
        },
    },
    'required': ['uuid'],
}
data = {
    'uuid': '123e4567-e89b-12d3-a456-426614174000',
}
ui.json_editor({'content': {'json': data}}, schema=schema)
```

<img width="907" height="315" alt="image"
src="https://github.com/user-attachments/assets/c197fc5f-24d4-40c6-adf9-cbdb4e1eb3ab"
/>




### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests
https://github.com/zauberzeug/nicegui/pull/5571/commits/e6fd20bfd8f363a9ccf39b7884dfe8a38b674b3f
- [x] Documentation
https://github.com/zauberzeug/nicegui/pull/5571/commits/0b00564eca4b67a983104d2f09108646e4f729ad

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>