Introduce on_component_click for ui.echart (#5578)
### Motivation

Fix #5576, where in the case of clicking indicator label, all we have in
`e.args` is key `componentType` and `name`, causing library code to
throw an error.

But #5577 no good because it shoehorns the component click into point
click handler, which has a risk of breaking downstream user code
expecteding to take a point from the handler args.

### Implementation

We have to segregate: 

- All mandatory keys available for point click? Go to `on_point_click`
handler with `EChartPointClickEventArguments`
- Any else case, go to `on_component_click` handler with
`EChartComponentClickEventArguments`

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>